supervisors remarks have been corrected
