+++ src/view/head/up-filters-view.js
@@ -1,31 +1,35 @@
-import { FilterLabeling } from '../../const';
+import { FilterLabelStartFrame } from '../../const';
 import { createElement } from '../../render';
 
 const createFiltersTemplate = () => {
-  const getSingleFilterDescription = (filterInfo, isChecked = false) => (
-    `<div class="trip-filters__filter">
+  const getSingleFilterDescription = (filterInfo) => {
+    const {lowCaseWord, capitalLetterWord, isChecked} = filterInfo;
+
+    return `<div class="trip-filters__filter">
 
     <input
-    id="filter-${filterInfo.lowCaseWord}"
+    id="filter-${lowCaseWord}"
     class="trip-filters__filter-input  visually-hidden"
     type="radio"
     name="trip-filter"
-    value="${filterInfo.lowCaseWord}"
+    value="${lowCaseWord}"
     ${isChecked ? 'checked' : ''}
     />
 
     <label
     class="trip-filters__filter-label"
     for="filter-everything"
-    >${filterInfo.capitalLetterWord}</label>
+    >${capitalLetterWord}</label>
 
-    </div>`
-  );
+    </div>`;
+  };
 
   return `<form class="trip-filters" action="#" method="get">
-    ${getSingleFilterDescription(FilterLabeling.EVERYTHING, true)}
-    ${getSingleFilterDescription(FilterLabeling.FUTURE)}
-    ${getSingleFilterDescription(FilterLabeling.PAST)}
+    ${Object
+    .values(FilterLabelStartFrame)
+    .map((StartFrame) => getSingleFilterDescription(StartFrame))
+    .join(' ')}
+
 
     <button class="visually-hidden" type="submit">Accept filter</button>
   </form>`;
